refactor: rename heartbeat publisher module
