+++ modelexpress_client/python/modelexpress/__init__.py
@@ -72,11 +72,10 @@ def register_modelexpress_loaders():
 from .client import MxClient  # noqa: F401
 from .gds_loader import MxGdsLoader  # noqa: F401
 from .gds_transfer import GdsTransferManager  # noqa: F401
-from .metadata.heartbeat import HeartbeatThread, PublisherThread  # noqa: F401
+from .metadata.publisher import PublisherThread  # noqa: F401
 
 __all__ = [
     "GdsTransferManager",
-    "HeartbeatThread",
     "MxClient",
     "MxGdsLoader",
     "PublisherThread",
